+++ packages/apps/stacc-website/sanity/plugins/website-template/src/caseStudyPage.js
@@ -2,10 +2,10 @@ import generatePageSchema from "./generatePageSchema"
 
 const page = generatePageSchema({
   name: "caseStudyPage",
-  title: "Case Study Page",
+  title: "Case Study Pages",
   allowSubpages: null,
   slugPrefix: "/clients/case-study/",
-  slugSource: "caseStudies[0].name",
+  slugSource: "name",
   additionalFields: [
     {
       type: "sectionHead",